Moved expression classes to `Expression` namespace
