+++ test/unit/Context/AbstractRightAssociativeExpressionTest.php
@@ -1,43 +1,45 @@
 <?php
 
-namespace Dhii\Expression\Test;
+namespace Dhii\Expression\Test\Expression;
 
-use Dhii\Expression\AbstractOperatorExpression;
+use Dhii\Expression\Expression\AbstractRightAssocOperatorExpression;
 use Dhii\Evaluable\EvaluableInterface;
 use Xpmock\TestCase;
 
 /**
- * Tests {@see Dhii\Expression\AbstractOperatorExpression}.
+ * Tests {@see Dhii\Expression\AbstractRightAssocOperatorExpression}.
  *
  * @since 0.1
  */
-class AbstractOperatorExpressionTest extends TestCase
+class AbstractRightAssociativeExpressionTest extends TestCase
 {
     /**
      * The name of the test subject.
      *
      * @since 0.1
      */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\\Expression\\AbstractOperatorExpression';
+    const TEST_SUBJECT_CLASSNAME = 'Dhii\\Expression\\Expression\\AbstractRightAssocOperatorExpression';
 
     /**
      * Creates an instance of the test subject.
      *
      * @since 0.1
      *
-     * @return AbstractOperatorExpression
+     * @return AbstractRightAssocOperatorExpression
      */
     public function createInstance()
     {
         $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
             ->_initBuffer(function () {
-                return 0;
+                return;
             })
             ->_defaultValue(function () {
                 return 0;
             })
             ->_operator(function ($left, $right) {
-                return intval($left) + intval($right);
+                return is_null($left)
+                    ? $right
+                    : $left / $right;
             })
         ;
 
@@ -82,10 +84,10 @@ class AbstractOperatorExpressionTest extends TestCase
     {
         $subject = $this->createInstance();
         $subject->this()->terms = array(
-            $this->mockTerm(5),
+            $this->mockTerm(2),
         );
 
-        $this->assertEquals(5, $subject->this()->_evaluate());
+        $this->assertEquals(2, $subject->this()->_evaluate());
     }
 
     /**
@@ -97,11 +99,11 @@ class AbstractOperatorExpressionTest extends TestCase
     {
         $subject = $this->createInstance();
         $subject->this()->terms = array(
-            $this->mockTerm(5),
-            $this->mockTerm(8),
+            $this->mockTerm(2),
+            $this->mockTerm(10),
         );
 
-        $this->assertEquals(13, $subject->this()->_evaluate());
+        $this->assertEquals(5, $subject->this()->_evaluate());
     }
 
     /**
@@ -113,13 +115,11 @@ class AbstractOperatorExpressionTest extends TestCase
     {
         $subject = $this->createInstance();
         $subject->this()->terms = array(
-            $this->mockTerm(5),
-            $this->mockTerm(8),
             $this->mockTerm(2),
+            $this->mockTerm(10),
             $this->mockTerm(4),
-            $this->mockTerm(3),
         );
 
-        $this->assertEquals(22, $subject->this()->_evaluate());
+        $this->assertEquals(0.2, $subject->this()->_evaluate());
     }
 }
